+++ packages/react-core/src/filters/tileFeatures.d.ts
@@ -1,5 +1,3 @@
 import { TileFeatures, TileFeaturesResponse } from '../types';
-import { Polygon, MultiPolygon } from 'geojson';
 
-export function getGeometryToIntersect(viewport: number[] | null, geometry: Polygon | MultiPolygon | null): Polygon | MultiPolygon | null;
 export function tileFeatures(arg: TileFeatures): TileFeaturesResponse;