+++ packages/react-core/src/index.js
@@ -13,6 +13,7 @@ export { debounce } from './utils/debounce';
 export { throttle } from './utils/throttle';
 export { randomString } from './utils/randomString';
 export { assert as _assert } from './utils/assert';
+export { getGeometryToIntersect, isGlobalViewport } from './utils/geo';
 
 export { makeIntervalComplete } from './utils/makeIntervalComplete';
 
@@ -34,7 +35,7 @@ export {
   applyFilters as _applyFilters
 } from './filters/Filter';
 
-export { tileFeatures, getGeometryToIntersect } from './filters/tileFeatures';
+export { tileFeatures } from './filters/tileFeatures';
 export { geojsonFeatures } from './filters/geojsonFeatures';
 
 export { GroupDateTypes } from './operations/constants/GroupDateTypes';
@@ -49,6 +50,7 @@ export {
 } from './utils/featureSelectionConstants';
 
 export {
+  Flags as _FeatureFlags,
   hasFlag as _hasFeatureFlag,
   setFlags as _setFeatureFlags,
   clearFlags as _clearFeatureFlags