+++ packages/react-widgets/src/widgets/TimeSeriesWidget.js
@@ -11,7 +11,9 @@ import {
 import {
   GroupDateTypes,
   AggregationTypes,
-  _FilterTypes as FilterTypes
+  _FilterTypes as FilterTypes,
+  _hasFeatureFlag,
+  _FeatureFlags
 } from '@carto/react-core';
 import { capitalize, Menu, MenuItem, SvgIcon } from '@mui/material';
 import { PropTypes } from 'prop-types';
@@ -123,7 +125,8 @@ function TimeSeriesWidget({
   const {
     data = [],
     isLoading,
-    warning
+    warning,
+    remoteCalculation
   } = useWidgetFetch(getTimeSeries, {
     id,
     dataSource,
@@ -135,7 +138,8 @@ function TimeSeriesWidget({
       operation
     },
     global,
-    onError
+    onError,
+    attemptRemoteCalculation: _hasFeatureFlag(_FeatureFlags.REMOTE_WIDGETS)
   });
 
   const handleTimeWindowUpdate = useCallback(
@@ -241,6 +245,7 @@ function TimeSeriesWidget({
           global={global}
           droppingFeaturesAlertProps={droppingFeaturesAlertProps}
           noDataAlertProps={noDataAlertProps}
+          showDroppingFeaturesAlert={!remoteCalculation}
         >
           {(!!data.length || isLoading) && (
             <TimeSeriesWidgetUI